Load fjord PAR data from a NetCDF file downloaded by FjordLight.

This functions will load into the R environment the data within one NetCDF file that has been downloaded via fl_DownloadFjord.

fl_LoadFjord(fjord, dirdata = NULL, TS = FALSE, verbose = FALSE)

Arguments

fjord

Expects a character vector for one of the 8 available fjords. See fl_ListFjords for the list of possible choices.

dirdata

The directory where the user would like to load the data from.

TS

The default, FALSE, will prevent this function from loading the monthly bottom PAR values. Instead it will load all global, annual, and monthly climatology data. Set TS = TRUE to load all data, but note that these may be extremely large.

verbose

TRUE or FALSE (default) to provide a more verbose output of the data loading process.

Value

Data are loaded in as a complex list format. Containing most of the data fields described in the documentation for fl_DownloadFjord.

Author

Bernard Gentili & Robert Schlegel

Examples

# Choose fjord
fjord_code <- "test"

# Load global, annual, and monthly climatologies
fjorddata <- fl_LoadFjord(fjord_code, dirdata = system.file("extdata", package = "FjordLight"))

# Load ALL data
fjorddata_full <- fl_LoadFjord(fjord_code,
                               dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)