Extract monthly bottom PAR data from a NetCDF file downloaded by FjordLight.

This functions will extract the monthly bottom PAR data stored within a NetCDF file downloaded via fl_DownloadFjord. Note that these data are very large. If one would prefer to work with the smaller annual or monthly climatology values, instead use flget_climatology. There are options for how the user would like to subset the data, which data format the data should be extracted to, and if the user would like to plot the data in the process.

flget_PARbottomMonthlyTS(
  fjord,
  month = NULL,
  year = NULL,
  mode = "raster",
  PLOT = FALSE
)

Arguments

fjord

Expects the object loaded via fl_LoadFjord. NB: when loading the data one must set the argument fl_LoadFjord(..., TS = TRUE). See examples below.

month

The monthly values to extract. Accepts one or many integer values from 3 to 10. If no values are provided, the default value of NULL will be passed to the function, telling it to load all available months of data (i.e. 3:10). This is used in combination with year to determine which monthly data to extract.

year

The years of data to extract. Currently accepts one or many integer values from 2003 to 2022. If no values are provided, the default value of NULL will be passed to the function, telling it to load all available years of data (i.e. currently 2003:2022). This is used in combination with month to determine which monthly data to extract.

mode

Determines the format of the data loaded into the R environment. The default "raster" will load the data as a raster format. The other option "3col" will load the data as a data.frame with three columns.

PLOT

Boolean argument (default = FALSE) that tells the function if the user would like the loaded data to be plotted or not.

Value

Depending on which arguments the user chooses, this function will return the chosen monthly bottom PAR data as a RasterStack (mode = "raster") or data.frame (mode = "3col"). The data.frame will contain the following columns:

longitude

degree decimals

latitude

degree decimals

PARbottom

mol photons m-2 d-1

Author

Bernard Gentili

Examples

# Download+load data
fjord_code <- "kong"
fl_DownloadFjord(fjord_code)
#> directory FjordLight.d created
#> ---> downloading fjord kong 
#> kong downloaded in directory FjordLight.d 
fjorddata <- fl_LoadFjord(fjord_code, TS = TRUE) # NB: TS = TRUE

# Years 2003 to 2004 - months July to August
mts <- flget_PARbottomMonthlyTS(fjorddata, month = 7:8, year = 2003:2004, PLOT = FALSE)
print(mts)
#> class      : RasterStack 
#> dimensions : 290, 321, 93090, 4  (nrow, ncol, ncell, nlayers)
#> resolution : 0.00462, 0.000883  (x, y)
#> extent     : 11.1738, 12.65682, 78.85438, 79.11045  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> names      : MonthlyPARbottom.2003.07, MonthlyPARbottom.2003.08, MonthlyPARbottom.2004.07, MonthlyPARbottom.2004.08 
#> min values :             4.123321e-40,             0.000000e+00,             0.000000e+00,             2.662467e-44 
#> max values :                 32.26019,                 19.93828,                 26.65807,                 18.68446 
#>