Extract geo-parameters from a NetCDF file downloaded by FjordLight.

Given a NetCDF file downloaded via fl_DownloadFjord, this function extracts the central longitude and latitude for that site, as well as the surface are in km^2 for the shallow zone (< 50 m deep) and coastal zone (< 200 m deep).

Usage

flget_geoparameters(fjord)

Arguments

fjord

Expects the object loaded via fl_LoadFjord.

Value

A named vector containing the following items:

site_average_longitude

The central longitude of the fjord [degree decimals]

site_average_latitude

The central latitude of the fjord [degree decimals]

AreaOfCoastalZone

The surface area of the grid cells with a depth of < 200 m [km^2]

AreaOfShallowZone

The surface area of the grid cells with a depth of < 50 m [km^2]

Author

Bernard Gentili

Examples

# Load data
fjord_code <- "test"
fjorddata <- fl_LoadFjord(fjord_code, dirdata = system.file("extdata", package = "FjordLight"))

# Load geo-parameters
geo_params <- flget_geoparameters(fjorddata)

# Convert to a data.frame if desired
geo_params_df <- t(as.data.frame(geo_params))