Plot p function data extracted from NetCDF files downloaded by FjordLight.

Internal function used by flget_Pfunction that plots data contained within the NetCDF files downloaded via fl_DownloadFjord.

Usage

flplot_Pfunction(irrLev, g, period, month, year, Main = NULL, add = add, ...)

Arguments

irrLev

The chosen irradiance thresholds (x-axis).

g

The percent of the fjord (y-axis) with a depth of 200 m or shallower receiving an irradiance values (e.g. surface PAR) at or above the corresponding threshold (x-axis).

period

The period being displayed. I.e. yearly or monthly.

month

If monthly climatology data, which month(s) is being displayed.

year

If yearly data, which year(s) is being displayed

Main

The title of the plot. The default NULL will provide a title 'Pfunction yyy'. Where yyy is the annual or monthly period chosen.

add

Boolean value of TRUE or FALSE telling the function to add this plot to the current plot in the R Plots panel.

...

Any other desired base R plotting functionality.

Value

A base R plot of the yearly or monthly p function data will be provided.

Author

Bernard Gentili