Download fjord PAR data as NetCDF files.

This functions queries the FTP server where the NetCDF files are stored. It will retrieve the one file that matches the name provided to it via the fjord argument. Note that these files can be multiple gigabytes in size.

fl_DownloadFjord(fjord, dirdata = NULL)

Arguments

fjord

Expects a character vector for one of the 8 available fjords. See fl_ListFjords for the list of possible choices.

dirdata

The directory where the user would like to download the data. Default is "FjordLight.d".

Value

The downloaded NetCDF file contains the following variables:

bathymetry

depth [m]

land

elevation [m]

area

PixelArea_km2 [m]

AreaOfCoastalZone

Surface of Sea floor with a depth of between 0 and 200 meters [km2]

etc...

Author

Bernard Gentili

Examples

# Choose a fjord
fjord_code <- "test"

# Download it
# NB: One should provide a permanent directory when downloading a file.
fl_DownloadFjord(fjord_code, dirdata = tempdir())
#> ---> downloading fjord test
#> test downloaded in directory /tmp/Rtmpy3ofIz